+++ noodles-bam/src/reader/unmapped_records.rs
@@ -1,6 +1,7 @@
 use std::io::{self, Read};
 
 use noodles_bgzf as bgzf;
+use noodles_sam::AlignmentRecord;
 
 use crate::Record;